+++ internal/xsql/conn.go
@@ -105,7 +105,7 @@ func (c *conn) isClosed() bool {
 	if atomic.LoadUint32(&c.closed) == 1 {
 		return true
 	}
-	if c.session.Status() != options.SessionReady.String() {
+	if c.session.Status() != table.SessionReady {
 		c.setClosed()
 		return true
 	}